+++ clevertap-pushtemplates/src/main/java/com/clevertap/android/pushtemplates/PushTemplateReceiver.java
@@ -30,6 +30,9 @@ import androidx.core.app.NotificationCompat;
 import androidx.core.app.NotificationCompat.Builder;
 import androidx.core.app.RemoteInput;
 import com.clevertap.android.pushtemplates.content.PendingIntentFactory;
+import com.clevertap.android.pushtemplates.media.GifDecoderImpl;
+import com.clevertap.android.pushtemplates.media.TemplateMediaManager;
+import com.clevertap.android.pushtemplates.media.TemplateRepository;
 import com.clevertap.android.sdk.CleverTapAPI;
 import com.clevertap.android.sdk.CleverTapInstanceConfig;
 import com.clevertap.android.sdk.Constants;
