+++ clevertap-pushtemplates/src/main/java/com/clevertap/android/pushtemplates/media/GifDrawableAdapter.kt
@@ -1,14 +1,33 @@
-package com.clevertap.android.pushtemplates
+package com.clevertap.android.pushtemplates.media
 
 import android.content.Context
+import android.graphics.Bitmap
+import com.clevertap.android.pushtemplates.PTHttpBitmapLoader
 import com.clevertap.android.sdk.CleverTapInstanceConfig
 import com.clevertap.android.sdk.Logger
 import com.clevertap.android.sdk.bitmap.BitmapDownloadRequest
 import com.clevertap.android.sdk.bitmap.HttpBitmapLoader
-import com.clevertap.android.sdk.bitmap.HttpBitmapLoader.getHttpBitmap
 import com.clevertap.android.sdk.network.DownloadedBitmap
-import com.clevertap.android.sdk.network.DownloadedBitmap.Status
 import com.clevertap.android.sdk.network.DownloadedBitmapFactory
+import pl.droidsonroids.gif.GifDrawable
+
+interface GifDrawableAdapter {
+    fun create(bytes: ByteArray): GifDrawable
+    fun getFrameCount(drawable: GifDrawable): Int
+    fun getFrameAt(drawable: GifDrawable, index: Int): Bitmap
+    fun getDuration(drawable: GifDrawable): Int
+}
+
+class GifDrawableAdapterImpl : GifDrawableAdapter {
+    override fun create(bytes: ByteArray) = GifDrawable(bytes)
+
+    override fun getFrameCount(drawable: GifDrawable) = drawable.numberOfFrames
+
+    override fun getFrameAt(drawable: GifDrawable, index: Int): Bitmap =
+        drawable.seekToFrameAndGet(index)
+
+    override fun getDuration(drawable: GifDrawable) = drawable.duration
+}
 
 internal class TemplateRepository(val context: Context, val config: CleverTapInstanceConfig?) {
 
@@ -19,7 +38,7 @@ internal class TemplateRepository(val context: Context, val config: CleverTapIns
     internal fun getBytes(url: String): DownloadedBitmap {
         if (url.isBlank()) {
             Logger.v("Cannot download GIF: URL is empty")
-            return DownloadedBitmapFactory.nullBitmapWithStatus(Status.NO_IMAGE)
+            return DownloadedBitmapFactory.nullBitmapWithStatus(DownloadedBitmap.Status.NO_IMAGE)
         }
 
         val request = BitmapDownloadRequest(
@@ -39,7 +58,7 @@ internal class TemplateRepository(val context: Context, val config: CleverTapIns
     internal fun getBitmap(url: String): DownloadedBitmap {
         if (url.isBlank()) {
             Logger.v("Cannot download Bitmap: URL is empty")
-            return DownloadedBitmapFactory.nullBitmapWithStatus(Status.NO_IMAGE)
+            return DownloadedBitmapFactory.nullBitmapWithStatus(DownloadedBitmap.Status.NO_IMAGE)
         }
 
         val request = BitmapDownloadRequest(
@@ -49,7 +68,7 @@ internal class TemplateRepository(val context: Context, val config: CleverTapIns
             null,
         )
 
-        return getHttpBitmap(
+        return HttpBitmapLoader.getHttpBitmap(
             HttpBitmapLoader.HttpBitmapOperation.DOWNLOAD_ANY_BITMAP,
             request
         )
